Move SwitcherItem into Switcher.kt
